fix: 完善快捷方式裁切代码规范
